requirements.txt
tensorflow
rdkit
ogb
python-dotenv